style: make outputs and transactions protocols look similar
