+++ apps/omg/config/config.exs
@@ -7,7 +7,8 @@ config :omg,
   ethereum_events_check_interval_ms: 500,
   coordinator_eth_height_check_interval_ms: 6_000,
   metrics_collection_interval: 60_000,
-  output_types_modules: %{<<1>> => OMG.Output.FungibleMoreVPToken}
+  output_types_modules: %{<<1>> => OMG.Output.FungibleMoreVPToken},
+  tx_types_modules: %{<<188, 97, 78>> => OMG.State.Transaction.Payment}
 
 config :omg, :eip_712_domain,
   name: "OMG Network",
